Disables logstash functionality by default
The default config assumed logstash was already installed, and also
assumed that the var `logstash_conf_dir` was defined. Renamed the var to
a reimann_ namedspaced version, set it to `/etc/logstash/conf.d`, and
disabled the logstash functionality by default. This is to avoid
failures when running the role when the defaults aren't modified.